+++ code/system.ui/react.common/src/ui.hooks/index.ts
@@ -1,1 +1,2 @@
 export { useFocus } from './Focus';
+export { useSizeObserver } from './Size';
