+++ code/system.ui/react.util/src/index.mts
@@ -3,5 +3,4 @@ export * from './types.mjs';
 export { FC } from './ui/FC/index.mjs';
 
 export { useClickOutside, useClickInside } from './ui.hooks/useClick.mjs';
-export { useSizeObserver } from './ui.hooks/useSizeObserver.mjs';
 export { useMouseState } from './ui.hooks/useMouseState.mjs';
